projects/huyang/core/src/lib/+state/huyang-header/header.facade.spec.ts
import { NgModule } from '@angular/core';
import { TestBed } from '@angular/core/testing';
import { readFirst } from '@nrwl/angular/testing';

import { EffectsModule } from '@ngrx/effects';
import { StoreModule, Store } from '@ngrx/store';

import { NxModule } from '@nrwl/angular';

import { HeaderEffects } from './header.effects';
import { HeaderFacade } from './header.facade';

import { headerQuery } from './header.selectors';
import { LoadHeader, HeaderLoaded } from './header.actions';
import { HeaderState, Entity, initialState, reducer } from './header.reducer';

interface TestSchema {
  header: HeaderState;
}

describe('HeaderFacade', () => {
  let facade: HeaderFacade;
  let store: Store<TestSchema>;
  let createHeader;

  beforeEach(() => {
    createHeader = (id: string, name = ''): Entity => ({
      id,
      name: name || `name-${id}`
    });
  });

  describe('used in NgModule', () => {
    beforeEach(() => {
      @NgModule({
        imports: [
          StoreModule.forFeature('header', reducer, { initialState }),
          EffectsModule.forFeature([HeaderEffects])
        ],
        providers: [HeaderFacade]
      })
      class CustomFeatureModule {}

      @NgModule({
        imports: [
          NxModule.forRoot(),
          StoreModule.forRoot({}),
          EffectsModule.forRoot([]),
          CustomFeatureModule
        ]
      })
      class RootModule {}
      TestBed.configureTestingModule({ imports: [RootModule] });

      store = TestBed.get(Store);
      facade = TestBed.get(HeaderFacade);
    });

    /**
     * The initially generated facade::loadAll() returns empty array
     */
    it('loadAll() should return empty list with loaded == true', async done => {
      try {
        let list = await readFirst(facade.allHeader$);
        let isLoaded = await readFirst(facade.loaded$);

        expect(list.length).toBe(0);
        expect(isLoaded).toBe(false);

        facade.loadAll();

        list = await readFirst(facade.allHeader$);
        isLoaded = await readFirst(facade.loaded$);

        expect(list.length).toBe(0);
        expect(isLoaded).toBe(true);

        done();
      } catch (err) {
        done.fail(err);
      }
    });

    /**
     * Use `HeaderLoaded` to manually submit list for state management
     */
    it('allHeader$ should return the loaded list; and loaded flag == true', async done => {
      try {
        let list = await readFirst(facade.allHeader$);
        let isLoaded = await readFirst(facade.loaded$);

        expect(list.length).toBe(0);
        expect(isLoaded).toBe(false);

        store.dispatch(
          new HeaderLoaded([createHeader('AAA'), createHeader('BBB')])
        );

        list = await readFirst(facade.allHeader$);
        isLoaded = await readFirst(facade.loaded$);

        expect(list.length).toBe(2);
        expect(isLoaded).toBe(true);

        done();
      } catch (err) {
        done.fail(err);
      }
    });
  });
});
